+++ HackerNewsSwiftUI/Modules/UI/Feature/NewsDetails/NewsDetailsView.swift
@@ -2,35 +2,27 @@
 //  NewsDetailsView.swift
 //  HackerNewsSwiftUI
 //
-//  Created by Arnav Maurya on 2024/02/04.
+//  Created by Dinakar Maurya on 2024/02/04.
 //  Copyright © 2024 Dinakar Prasad Maurya. All rights reserved.
 //
 
 import Foundation
-import WebKit
 import SwiftUI
 
-struct NewsDetailsView: UIViewRepresentable {
-    
+struct NewsDetailsView: View {
     let urlString: String?
-    
-    func makeUIView(context: Context) -> WKWebView {
-        return WKWebView()
-    }
-    
-    func updateUIView(_ uiView: WKWebView, context: Context) {
-        if let safeString = urlString {
-            if let url = URL(string: safeString) {
-                let request = URLRequest(url: url)
-                uiView.load(request)
+
+    var body: some View {
+        VStack {
+            Text(urlString ?? "")
+            if let urlString = urlString, let url = URL(string: urlString) {
+                SharedWebView(url: url)
+                    .padding(.top, 0)
+                    .navigationBarTitle("New Details")
+                    .navigationBarTitleDisplayMode(.inline)
+            } else {
+                Text("Invalid URL")
             }
         }
-    }
-    
-}
-
-struct WebView_Previews: PreviewProvider {
-    static var previews: some View {
-        NewsDetailsView(urlString: "www.google.com")
     }
 }
